Upvalues moved out of object.c/.h
